backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
requests==2.26.0
python-dotenv==0.19.0
moviepy==1.0.3
PyJWT==2.8.0 